explicitly use int for all open enums except NotificationTypes
otherwise casting out-of-range values is undefined. enum_extensibility
only allows casting unlisted values into the existing range, not values
outside the range. e.g. for enum E{A=1, B=10}, enum_extensibility(open)
allows (E)5 but not (E)11. Civ uses enum E{A=-1}, so (E)-1 and (E)0 are
valid but nothing else, which is not what we want.

int is the safest type because MSVC already uses it, so it's least
likely to cause issues. we can narrow the type later if needed